carga de datos del csv, queda hacer el h)

+++ Integrador3/src/main/java/org/example/integrador3/service/dto/estudiante/response/EstudianteResponseDTO.java
@@ -2,7 +2,7 @@ package org.example.integrador3.service.dto.estudiante.response;
 
 import org.example.integrador3.domain.Estudiante;
 
-public record EstudianteResponseDTO(String nombre, String apellido, int dni, int numLibreta){
+public record EstudianteResponseDTO(String nombre, String apellido, long dni, int numLibreta){
     public EstudianteResponseDTO(Estudiante e) {
         this(e.getNombre(), e.getApellido(), e.getDocumento(), e.getNumLibreta());
     }
